fix(gateway): preserve protocol and type safety guards
